+++ kamehouse-shell/bin/kamehouse/docker/release/java11-release/java11-release-kamehouse-docker-keep-alive.sh
@@ -3,6 +3,7 @@
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing kamehouse-functions.sh" ; exit 99 ; fi
 
+importKamehouse common/functions/keep-alive/keep-alive-functions.sh
 importKamehouse common/functions/kamehouse/docker-functions.sh
 importKamehouse common/functions/kamehouse/docker/release/java11-release-functions.sh
 
@@ -16,8 +17,7 @@ initScriptEnv() {
   SERVICE_ARGS=""
 }
 
-mainProcess() {
-  checkKeepAliveScriptsEnabled
+runKeepAlive() {
   log.warn "User running this script needs ${COL_RED}sudo netstat${COL_DEFAULT_LOG} permissions"
   PID=`sudo netstat -nltp | grep ":${DOCKER_HTTP_PORT} " | awk '{print $7}' | cut -d '/' -f 1`
   if [ -z "${PID}" ]; then