+++ kamehouse-shell/bin/lin/keep-alive/keep-alive-kamehouse-docker.sh
@@ -3,6 +3,7 @@
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing kamehouse-functions.sh" ; exit 99 ; fi
 
+importKamehouse common/functions/keep-alive/keep-alive-functions.sh
 importKamehouse common/functions/kamehouse/docker-functions.sh
 
 initScriptEnv() {
@@ -13,8 +14,7 @@ initScriptEnv() {
   DOCKER_PROFILE="${DEFAULT_DOCKER_PROFILE}"
 }
 
-mainProcess() {
-  checkKeepAliveScriptsEnabled
+runKeepAlive() {
   log.warn "User running this script needs ${COL_RED}sudo netstat${COL_DEFAULT_LOG} permissions"
   PID=`sudo netstat -nltp | grep ":${DOCKER_PORT_HTTP} " | awk '{print $7}' | cut -d '/' -f 1`
   if [ -z "${PID}" ]; then